Give a released code slot to later code, and check an id's generation
Since #338 the VM lets go of a finished form's code, but VmState::code_store
kept its slot: ids came from a process-wide counter, dense and never reused,
and the store was indexed by them. Memory still grew by about 8 bytes a form,
to 41.5 MB at 3,200,000 forms streamed on standard input, and the u32 counter
would wrap after about four billion code objects and let ids collide.

An id is now a slot in one VmState's store and the generation of the code in
it, packed in 64 bits. The compiler names code by a label, which the
MakeClosure in the code around it uses too; load_unit gives each code object a
slot and replaces both. A slot let go is given to the next code loaded, with
the next generation, so the store grows with the code loaded at once. A slot
that has used every generation is not given out again, so ids never repeat.

A lookup checks that the slot's code has the id asked for, so an id outliving
its code finds nothing rather than running whatever took its slot. An empty
slot holds a shared placeholder whose id is a label rather than None, so that
one comparison turns away an empty slot too: with an Option slot the check
cost 2.6% of a call-heavy loop, and this way 1.4%.

The same 3,200,000 forms now peak at 15 MB, as 200,000 do. Neither an
Instruction nor a HeapObjectData grows with the wider id.

Fixes #352

+++ crates/patina-core/src/heap/mod.rs
@@ -200,8 +200,9 @@ pub enum HeapObjectData {
     Ephemeron(RefCell<Option<(TaggedValue, TaggedValue)>>),
     /// A VM bytecode closure: code id + captured free variables + globals env.
     VmClosure {
-        /// Serialised as `u32` to avoid a direct dependency on patina-vm types.
-        code_id: u32,
+        /// The VM's `CodeObjectId`, as its raw `u64`, to avoid a direct
+        /// dependency on patina-vm types.
+        code_id: u64,
         free_vars: Vec<TaggedValue>,
         /// The global environment this closure was compiled against.
         /// `LoadGlobal`/`StoreGlobal` use this instead of `VmState::globals`.
@@ -388,7 +389,7 @@ pub struct Heap {
     /// that nothing can run again (#338). A closure holds its code by id
     /// alone, so this is the only way its death reaches that count. `None`
     /// (the default) means nobody consumes them and sweep records nothing.
-    gc_freed_closure_code_ids: Option<Vec<u32>>,
+    gc_freed_closure_code_ids: Option<Vec<u64>>,
 
     /// Nesting depth of scopes that hold live values unreachable from any
     /// registered root (nested trampolines, library-body loops). Collection
@@ -598,7 +599,7 @@ impl Heap {
 
     /// Drain the code ids of the VM closures reclaimed since the last drain,
     /// one entry per closure.
-    pub fn take_gc_freed_closure_code_ids(&mut self) -> Vec<u32> {
+    pub fn take_gc_freed_closure_code_ids(&mut self) -> Vec<u64> {
         self.gc_freed_closure_code_ids
             .as_mut()
             .map(std::mem::take)
@@ -1238,11 +1239,11 @@ impl Heap {
 
     /// Allocate a VM closure on the heap.
     ///
-    /// `code_id` is a `u32` (the raw `CodeObjectId(u32)` value) to avoid
+    /// `code_id` is a `u64` (the raw `CodeObjectId(u64)` value) to avoid
     /// a direct crate dependency on `patina-vm`.
     pub fn alloc_vm_closure(
         &mut self,
-        code_id: u32,
+        code_id: u64,
         free_vars: Vec<TaggedValue>,
         globals: Rc<crate::environment::Environment>,
     ) -> TaggedValue {
@@ -1256,7 +1257,7 @@ impl Heap {
     /// Retrieve the `(code_id, free_vars)` pair from a VM closure pointer.
     ///
     /// Returns `None` if `val` is not a `VmClosure` object.
-    pub fn get_vm_closure(&self, val: TaggedValue) -> Option<(u32, Vec<TaggedValue>)> {
+    pub fn get_vm_closure(&self, val: TaggedValue) -> Option<(u64, Vec<TaggedValue>)> {
         if !val.is_object() {
             return None;
         }
@@ -1272,7 +1273,7 @@ impl Heap {
     /// free variables. This is the call-path accessor — `get_vm_closure`
     /// clones the whole `free_vars` vector, which callers that just need to
     /// dispatch the closure would immediately discard.
-    pub fn get_vm_closure_code_id(&self, val: TaggedValue) -> Option<u32> {
+    pub fn get_vm_closure_code_id(&self, val: TaggedValue) -> Option<u64> {
         if !val.is_object() {
             return None;
         }
